sauv avant weekend

+++ Fdf/fdf_v4/srcs/get_tram_elem.c
@@ -6,7 +6,7 @@
 /*   By: mbocquel <mbocquel@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/13 17:09:43 by mbocquel          #+#    #+#             */
-/*   Updated: 2022/12/16 21:24:41 by mbocquel         ###   ########.fr       */
+/*   Updated: 2022/12/16 22:59:14 by mbocquel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,8 @@ t_trame	*get_tram_elem(int line, int col, t_trame *begin)
 	elem = begin;
 	while (elem)
 	{
-		if (elem->point_3d.y_3d == line && elem->point_3d.x_3d == col)
+		if (elem->point_3d.y_3d == (float)line
+			&& elem->point_3d.x_3d == (float)col)
 			return (elem);
 		else
 			elem = elem->next;
@@ -34,7 +35,8 @@ t_trame	*get_tram_elem_from_end(int line, int col, t_trame *begin)
 	elem = begin;
 	while (elem)
 	{
-		if (elem->point_3d.y_3d == line && elem->point_3d.x_3d == col)
+		if (elem->point_3d.y_3d == (float)line
+			&& elem->point_3d.x_3d == (float)col)
 			return (elem);
 		else
 			elem = elem->previous;
