+++ Fdf/fdf_v4/srcs/prod_mat.c
@@ -6,7 +6,7 @@
 /*   By: mbocquel <mbocquel@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/13 17:16:12 by mbocquel          #+#    #+#             */
-/*   Updated: 2022/12/16 21:51:02 by mbocquel         ###   ########.fr       */
+/*   Updated: 2022/12/16 23:00:11 by mbocquel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,16 +16,17 @@ t_point2d_fl	prod_mat3x2_p3d(t_mat3x2 mat, t_point_3d point_3d)
 {
 	t_point2d_fl	point;
 
-	point.xf = mat.l1[0] * (float)point_3d.x_3d + mat.l1[1]
-		* (float)point_3d.y_3d + mat.l1[2] * (float)point_3d.z_3d;
-	point.yf = mat.l2[0] * (float)point_3d.x_3d + mat.l2[1]
-		* (float)point_3d.y_3d + mat.l2[2] * (float)point_3d.z_3d;
+	point.xf = mat.l1[0] * point_3d.x_3d + mat.l1[1]
+		* point_3d.y_3d + mat.l1[2] * point_3d.z_3d;
+	point.yf = mat.l2[0] * point_3d.x_3d + mat.l2[1]
+		* point_3d.y_3d + mat.l2[2] * point_3d.z_3d;
 	return (point);
 }
 
 t_point_3d	prod_mat3x3_p3d(t_mat3x3 mat, t_point_3d p_3d)
 {
 	t_point_3d	result;
+
 	result.x_3d = mat.l1[0] * p_3d.x_3d + mat.l1[1] * p_3d.y_3d
 		+ mat.l1[2] * p_3d.z_3d;
 	result.y_3d = mat.l2[0] * p_3d.x_3d + mat.l2[1] * p_3d.y_3d
